build quiz api

+++ src/main/java/phucitdev/course/modules/snap_lessonquiz/service/Impl/SnapLessonQuizServiceImpl.java
@@ -1,7 +1,10 @@
 package phucitdev.course.modules.snap_lessonquiz.service.Impl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import phucitdev.course.commo.exception.auth.BadRequestException;
 import phucitdev.course.commo.exception.classroom.NotFoundException;
+import phucitdev.course.modules.lesson_quiz.entity.StudentQuizSubmission;
+import phucitdev.course.modules.lesson_quiz.repository.StudentQuizSubmissionRepository;
 import phucitdev.course.modules.snap_lessonquiz.dto.DeleteSnapLessonQuizResponse;
 import phucitdev.course.modules.snap_lessonquiz.entity.SnapLessonQuiz;
 import phucitdev.course.modules.snap_lessonquiz.repository.SnapLessonQuizRepository;
@@ -11,11 +14,17 @@ import java.util.UUID;
 public class SnapLessonQuizServiceImpl implements SnapLessonQuizService {
      @Autowired
     SnapLessonQuizRepository  snapLessonQuizRepository;
+     @Autowired
+     StudentQuizSubmissionRepository studentQuizSubmissionRepository;
     @Override
     public DeleteSnapLessonQuizResponse deleteSnapLessonQuiz(UUID snapLessonQuizId) {
       SnapLessonQuiz snapLessonQuiz = snapLessonQuizRepository.findSnapLessonQuizById(snapLessonQuizId)
                 .orElseThrow(() -> new NotFoundException("SnapLessonQuiz không tồn tại"));
+        boolean hasSubmission = studentQuizSubmissionRepository.existsBySnapLessonQuizId(snapLessonQuizId);
+        if (hasSubmission) {
+            throw new BadRequestException("Không thể gỡ vì đã có học viên nộp bài");
+        }
         snapLessonQuizRepository.delete(snapLessonQuiz);
-        return new DeleteSnapLessonQuizResponse("SnapLessonQuiz đã bị xoá!");
+        return new DeleteSnapLessonQuizResponse("SnapLessonQuiz đã bị gỡ!");
     }
 }
